feat: changes to database scheme

+++ BookingService/BookingService.Application/DTOs/WalletTopUpDto.cs
@@ -2,6 +2,6 @@
 {
     public record WalletTopUpDto
     {
-        public double Amount { get; init; }
+        public decimal Amount { get; init; }
     }
 }
